serial/linux: Fix TTY opening and closing logic
The serial port was being opened with the wrong parameters, such as the baud rate not using the rate macros but supplying the rate as an integer directly instead which isn't valid. There were also issues with the serial port not being fully reset (flushed/drained) before setting all the configuration options, and the port wasn't drained on destruction which is necessary to ensure that all pending bytes are sent out.