serial/linux: Use fstat to detect serial device disconnection
The validity of the FD was checked prior, which turned out to not be enough as the FD could still be valid but the device disconnected. This change adds a check for the device disconnection by using fstat to check if the file has any hard links left which is far more reliable.