serial/linux: Add serial port path to friendly name
The friendly name of the serial device was retrieved from sysfs, but this would lead to duplicates if there were multiple devices of the same type. To allow for differentiation between them, and to match the Windows friendly name, the path to the device is now included in the friendly name.